move all specs files to tests dir
