+++ src/tests/__fixtures__/preparedData.js
@@ -1,7 +1,7 @@
 import { List } from 'immutable';
 
 import networkDataMock from './network';
-import { prepareViewerData } from '../utils';
+import { prepareViewerData } from '../../utils';
 
 const preparedMockData = new List(prepareViewerData(networkDataMock.log.entries).data);
