+++ src/tests/__tests__/App.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import App from '../App';
+import App from '../../App';
 
 it('renders without crashing', () => {
   shallow(<App />);
