+++ src/tests/__tests__/state/reducer.spec.js
@@ -1,7 +1,7 @@
 import { Map } from 'immutable';
 
-import * as types from '../../state/network/types';
-import { reducer, initialState } from '../../state/network/reducer';
+import * as types from '../../../state/network/types';
+import { reducer, initialState } from '../../../state/network/reducer';
 import networkDataMock from '../../__fixtures__/network.json';
 import preparedMockData from '../../__fixtures__/preparedData';
